dom/tests/mochitest/dom-level0/test_location.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for location object behaviors</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<iframe id="ifr" onload="runTest()"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var count = 0;
var firstlocation;
var lastlocation;

function runTest() {
  ++count;
  if (count == 1) {
    firstlocation = $('ifr').contentWindow.location;
    firstlocation.existingprop = 'fail';
    firstlocation.href = 'file_location.html';
    return;
  }

  if (count == 2) {
    lastlocation = $('ifr').contentWindow.location;
    is(lastlocation.iframeprop, 42, 'can read the new prop');
    ok(firstlocation !== lastlocation, 'got a new location object');
    // firstlocation should still work.
    firstlocation.href = 'http://example.com/tests/dom/tests/mochitest/dom-level0/file_location.html';
    return;
  }

  if (count == 3) {
    try {
      isnot($('ifr').contentWindow.location.iframeprop, 42, "shouldn't see this");
    } catch (e) {
      ok(/Permission denied/.test(e), "correctly threw a security error");
    }

    // XXX this doesn't work - file a bug on me!
    //firstlocation.href = 'http://mochi.test:8888/tests/dom/tests/mochitest/dom-level0/file_location.html';
    todo(false, 'can use old same-origin location object to set cross-origin frame');
    $('ifr').contentWindow.location = 'file_location.html';
    return;
  }

  is(lastlocation.iframeprop, 42, 'can still read old values of the location object');
  ok(lastlocation !== $('ifr').contentWindow.location, 'location objects are distinct');

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
